[Refactor]코드 추상화 및 정리
