+++ app/(tabs)/const/beverageDataSet.ts
@@ -1,0 +1,11 @@
+import { BeverageVariantType } from "@/components/CustomSlide/CustomSlide.types";
+
+const beverageDataSet: (BeverageVariantType | "addYourself")[] = [
+  "water",
+  "coffee",
+  "greenTea",
+  "cola",
+  "addYourself",
+];
+
+export default beverageDataSet;
